Refactor TestPlan manifest init into its own file
